change a little

+++ cpp/3d.h
@@ -10,21 +10,21 @@
 #include <vector>
 
 struct tdDrawPolygon {
-    std::array<long double,3> p1;
-    std::array<long double,3> p2;
-    std::array<long double,3> p3;
+    std::array<double,3> p1;
+    std::array<double,3> p2;
+    std::array<double,3> p3;
     std::array<unsigned char,3> color;
-    long double distance = 0;
+    double distance = 0;
 };
 typedef std::vector<struct tdDrawPolygon> tdDrawObject;
 unsigned char* iarr;
-const long double pi = 3.14159;
+const double pi = 3.14159;
 
 class tdDraw {
     public:
-        std::array<long double,3> campos = {0,0,0};
-        std::array<long double,2> camangle = {0,0};
-        long double trifv[4];
+        std::array<double,3> campos = {0,0,0};
+        std::array<double,2> camangle = {0,0};
+        double trifv[4];
         int display[2];
         tdDrawObject obj;
         int frame;
@@ -38,7 +38,7 @@ class tdDraw {
         void setObj(tdDrawObject object) {
             obj = object;
         }
-        void setCamera(std::array<long double,3> new_campos,std::array<long double,2> new_camangle) {
+        void setCamera(std::array<double,3> new_campos,std::array<double,2> new_camangle) {
             campos = new_campos;
             camangle = new_camangle;
         }
@@ -48,10 +48,10 @@ class tdDraw {
             frame++;
 		    std::cout << "f " << frame << std::endl;
             int x = width;int y = height;
-            long double maxlen = 100;
+            double maxlen = 100;
 
             iarr = new uchar [x*y*4];
-            long double* zbuf = new long double [x*y];
+            double* zbuf = new double [x*y];
             for (int i=0;i<x*y;i++) {zbuf[i]=maxlen;iarr[i*4+2]=sky[0];iarr[i*4+1]=sky[1];iarr[i*4+0]=sky[2];}
 
             tda vl = VNormalized({30,10,20});
@@ -73,36 +73,36 @@ class tdDraw {
                 tda v13 = {t.p3[0]-t.p1[0],t.p3[1]-t.p1[1],t.p3[2]-t.p1[2]};
 
                 tda normal = VNormalized(VCProduct(v12,v13));
-                long double angl = VIProduct(vl,normal);
-                long double light = 0;
+                double angl = VIProduct(vl,normal);
+                double light = 0;
 
-                int xmax = (int)std::min(std::max(std::max((long double)p1[0],(long double)p2[0]),(long double)p3[0]),(long double)display[0]);
-                int xmin = (int)std::max(std::min(std::min((long double)p1[0],(long double)p2[0]),(long double)p3[0]),(long double)0);
-                int ymax = (int)std::min(std::max(std::max((long double)p1[1],(long double)p2[1]),(long double)p3[1]),(long double)display[1]);
-                int ymin = (int)std::max(std::min(std::min((long double)p1[1],(long double)p2[1]),(long double)p3[1]),(long double)0);
+                int xmax = (int)std::min(std::max(std::max((double)p1[0],(double)p2[0]),(double)p3[0]),(double)display[0]);
+                int xmin = (int)std::max(std::min(std::min((double)p1[0],(double)p2[0]),(double)p3[0]),(double)0);
+                int ymax = (int)std::min(std::max(std::max((double)p1[1],(double)p2[1]),(double)p3[1]),(double)display[1]);
+                int ymin = (int)std::max(std::min(std::min((double)p1[1],(double)p2[1]),(double)p3[1]),(double)0);
 
-                long double al = length3d({campos,t.p1});
-                long double bl = length3d({campos,t.p2});
-                long double cl = length3d({campos,t.p3});
-                long double ap = p1[2];
-                long double bp = p2[2];
-                long double cp = p3[2];
+                double al = length3d({campos,t.p1});
+                double bl = length3d({campos,t.p2});
+                double cl = length3d({campos,t.p3});
+                double ap = p1[2];
+                double bp = p2[2];
+                double cp = p3[2];
                 for (int iy=ymin;iy<ymax;iy++) {
                     for (int ix=xmin;ix<xmax;ix++) {
                         int idex = (height-iy)*x+ix;
-                        if (inclusion((t2da){(long double)ix,(long double)iy},{p1,p2,p3})) {
-                            tda td = is_p({p1,(tda){(long double)ix,(long double)iy,0.0}},{p2,p3});
+                        if (inclusion((t2da){(double)ix,(double)iy},{p1,p2,p3})) {
+                            tda td = is_p({p1,(tda){(double)ix,(double)iy,0.0}},{p2,p3});
 
-                            long double dl = bl+(length2d({p2,td})/length2d({p2,p3}))*(cl-bl);
-                            long double pl = al+(length2d({p1,{(long double)ix,(long double)iy,0.0}})/length2d({p1,td}))*(dl-al);
+                            double dl = bl+(length2d({p2,td})/length2d({p2,p3}))*(cl-bl);
+                            double pl = al+(length2d({p1,{(double)ix,(double)iy,0.0}})/length2d({p1,td}))*(dl-al);
 
-                            long double dp = bp+(length2d({p2,td})/length2d({p2,p3}))*(cp-bp);
-                            long double pp = ap+(length2d({p1,{(long double)ix,(long double)iy,0.0}})/length2d({p1,td}))*(dp-ap);
+                            double dp = bp+(length2d({p2,td})/length2d({p2,p3}))*(cp-bp);
+                            double pp = ap+(length2d({p1,{(double)ix,(double)iy,0.0}})/length2d({p1,td}))*(dp-ap);
 
                             if (pp>0&&zbuf[idex]>pl) {
                                 zbuf[idex] = pl;
                                 int index = idex*4;
-                                long double light = (std::max(angl,angl*0.1)*0.9+0.3)*(1000/(std::pow(pl,2)+1000));
+                                double light = (std::max(angl,angl*0.1)*0.9+0.3)*(1000/(std::pow(pl,2)+1000));
                                 iarr[index+2] = t.color[0]*light;
                                 iarr[index+1] = t.color[1]*light;
                                 iarr[index+0] = t.color[2]*light;
@@ -116,15 +116,15 @@ class tdDraw {
             return iarr;
         }
     private:
-        typedef std::array<long double,3> tda;
-        typedef std::array<long double,2> t2da;
+        typedef std::array<double,3> tda;
+        typedef std::array<double,2> t2da;
         typedef unsigned char uchar;
         void sortPolygon() {
         }
         tda pos_3t2d(tda p) {
             tda p1 = rotate3d_x(rotate3d_z({p[0]-campos[0],p[1]-campos[1],p[2]-campos[2]}));
-            long double l = std::abs(14/p1[1]);
-            long double s = display[0]/40;
+            double l = std::abs(14/p1[1]);
+            double s = display[0]/40;
             return {std::floor(p1[0]*s*l+display[0]/2),std::floor(-p1[2]*s*l+display[1]/2),p1[1]};
         }
         void sortP() {
@@ -135,43 +135,43 @@ class tdDraw {
         }
         tda is_p(std::array<tda,2> l1,std::array<tda,2> l2) {
             // 参考 https://mf-atelier.sakura.ne.jp/mf-atelier2/a1/
-            long double div = (l1[1][1]-l1[0][1])*(l2[1][0]-l2[0][0])-(l1[1][0]-l1[0][0])*(l2[1][1]-l2[0][1]);
+            double div = (l1[1][1]-l1[0][1])*(l2[1][0]-l2[0][0])-(l1[1][0]-l1[0][0])*(l2[1][1]-l2[0][1]);
             if (div==0) {return {NAN,NAN,0};};
-            long double d1 = l2[0][1]*l2[1][0]-l2[0][0]*l2[1][1];
-            long double d2 = l1[0][1]*l1[1][0]-l1[0][0]*l1[1][1];
-            long double x = (d1*(l1[1][0]-l1[0][0])-d2*(l2[1][0]-l2[0][0]))/div;
-            long double y = (d1*(l1[1][1]-l1[0][1])-d2*(l2[1][1]-l2[0][1]))/div;
+            double d1 = l2[0][1]*l2[1][0]-l2[0][0]*l2[1][1];
+            double d2 = l1[0][1]*l1[1][0]-l1[0][0]*l1[1][1];
+            double x = (d1*(l1[1][0]-l1[0][0])-d2*(l2[1][0]-l2[0][0]))/div;
+            double y = (d1*(l1[1][1]-l1[0][1])-d2*(l2[1][1]-l2[0][1]))/div;
             return {x,y,0};
         }
-        long double squared_length3d(std::array<tda,2> p) {
+        double squared_length3d(std::array<tda,2> p) {
             return std::pow(p[0][0]-p[1][0],2)+std::pow(p[0][1]-p[1][1],2)+std::pow(p[0][2]-p[1][2],2);
         }
-        long double length3d(std::array<tda,2> p) {
+        double length3d(std::array<tda,2> p) {
             return std::sqrt(std::pow(p[0][0]-p[1][0],2)+std::pow(p[0][1]-p[1][1],2)+std::pow(p[0][2]-p[1][2],2));
         }
-        long double length2d(std::array<tda,2> p) {
+        double length2d(std::array<tda,2> p) {
             return std::sqrt(std::pow(p[0][0]-p[1][0],2)+std::pow(p[0][1]-p[1][1],2));
         }
         tda VCProduct(tda a,tda b) {
             return {a[1]*b[2]-a[2]*b[1],a[2]*b[0]-a[0]*b[2],a[0]*b[1]-a[1]*b[0]};
         }
-        long double VIProduct(tda a,tda b) {
+        double VIProduct(tda a,tda b) {
             return a[0]*b[0]+a[1]*b[1]+a[2]*b[2];
         }
         tda VMM3(tda a) {
             return {-a[0],-a[1],-a[2]};
         }
         tda VNormalized(tda a) {
-            long double l = length3d({(tda){0,0,0},(tda)a});
+            double l = length3d({(tda){0,0,0},(tda)a});
             return {a[0]/l,a[1]/l,a[2]/l};
         }
         bool inclusion(t2da p1,std::array<tda,3> carr) {
             t2da a = {carr[0][0]-p1[0],carr[0][1]-p1[1]};
             t2da b = {carr[1][0]-p1[0],carr[1][1]-p1[1]};
             t2da c = {carr[2][0]-p1[0],carr[2][1]-p1[1]};
-            long double ab = a[0]*b[1]-a[1]*b[0];
-            long double bc = b[0]*c[1]-b[1]*c[0];
-            long double ca = c[0]*a[1]-c[1]*a[0];
+            double ab = a[0]*b[1]-a[1]*b[0];
+            double bc = b[0]*c[1]-b[1]*c[0];
+            double ca = c[0]*a[1]-c[1]*a[0];
             return ab<=0&&bc<=0&&ca<=0;
         }
         tda rotate3d_x(tda p) {
